requirements.txt
gspread==5.8.0
pyOpenSSL==23.1.1
pyOpenSSL==23.1.1
PyYAML==6.0
PyYAML==6.0
Requests==2.30.0
